pyproject.toml
[project]
name = "hospital-emergency-env"
version = "0.1.0"
description = "Hospital Emergency Resource Allocation OpenEnv Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "openai>=2.7.2",
    "httpx>=0.28.1",
    "python-multipart>=0.0.9",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]